factor out a train_expert script
