+++ scripts/arguments.py
@@ -1,18 +1,17 @@
-from . import utils, environments, training, policies, irl
-from baselines.ppo2.policies import MlpPolicy, CnnPolicy
+#from atari_irl import utils, environments, training, policies, irl
+
 import os.path as osp
 import os
 import argparse
-import tensorflow as tf
-from baselines import logger
-from sandbox.rocky.tf.envs.base import TfEnv
-import pickle
+#import tensorflow as tf
+#from baselines import logger
+#from sandbox.rocky.tf.envs.base import TfEnv
+#import pickle
 
+EXPERT_POLICY_FILENAME_ARG = '--expert_file'
 
-def atari_arg_parser(parser=None):
+def atari_arg_parser(parser):
     # see baselines.common.cmd_util
-    if not parser:
-        parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     parser.add_argument('--env', help='environment ID', default='PongNoFrameskip-v0')
     parser.add_argument('--seed', help='RNG seed', type=int, default=0)
     parser.add_argument('--num_timesteps', type=int, default=int(10e6))
@@ -42,43 +41,6 @@ def add_irl_args(parser):
         help='filename for the IRL policy',
         default='irl_policy_params.pkl'
     )
-
-
-def train_expert(args):
-    utils.logger.configure()
-    with utils.TfContext():
-        with utils.EnvironmentContext(
-                env_name=args.env,
-                n_envs=args.num_envs,
-                seed=args.seed,
-                **environments.atari_modifiers
-        ) as context:
-            learner = training.Learner(
-                CnnPolicy, context.environments,
-                total_timesteps=args.num_timesteps,
-                vf_coef=0.5, ent_coef=0.01,
-                nsteps=128, noptepochs=4, nminibatches=4,
-                gamma=0.99, lam=0.95,
-                lr=lambda alpha: alpha * 2.5e-4,
-                cliprange=lambda alpha: alpha * 0.1
-            )
-
-            for policy, update, mean_reward in learner.learn_and_yield(
-                    lambda l: (l.policy, l.update, l.mean_reward), 100,
-                    log_freq=1
-            ):
-                checkdir = osp.join(utils.logger.get_dir(), 'checkpoints')
-                os.makedirs(checkdir, exist_ok=True)
-                savepath = osp.join(checkdir, 'update-{}'.format(update))
-                print('Saving to', savepath)
-                policy.save(
-                    savepath,
-                    mean_reward=learner.mean_reward,
-                    update=update,
-                    seed=args.seed
-                )
-
-            policies.run_policy(model=policy.model, environments=policy.envs)
 
 
 def generate_trajectories(args):
@@ -153,7 +115,4 @@ def run_irl(args):
             policy.show_run_in_gym_env(context.environments)
 
 
-if __name__ == '__main__':
-    parser = atari_arg_parser()
-    args = parser.parse_args()
-    train_expert(args)
+
